+++ src/__stories__/__data__/other/index.ts
@@ -3,5 +3,6 @@ export * from './crosshair';
 export * from './legend-position';
 export * from './line-and-bar';
 export * from './lines';
+export * from './shapes';
 export * from './tooltip';
 export * from './legend';